backend/requirements.txt
Django
djangorestframework
corsheaders
djangorestframework-simplejwt
google-generativeai
pydantic>=2.0.0